+++ spec/policies/result_policy_spec.rb
@@ -24,11 +24,70 @@ describe ResultPolicy do
       let(:role) { create :student }
       let(:record) { create :complete_result }
     end
-    succeed 'role is a student who is part of the grouping' do
+    describe 'role is a student who is part of the grouping' do
       let(:role) { create :student }
       let(:grouping) { create :grouping_with_inviter_and_submission, inviter: role }
       let(:record) { grouping.current_result }
       let(:assignment) { record.grouping.assignment }
+      succeed 'assignment.release_with_urls is false' do
+        before { assignment.update! release_with_urls: false }
+      end
+      context 'assignment.release_with_urls is true' do
+        before { assignment.update! release_with_urls: true }
+        let(:context) { { role: role, real_user: role.user, view_token: view_token } }
+        failed 'the view token does not match the record token' do
+          let(:view_token) { "#{record.view_token}abc123" }
+        end
+        context 'the view token matches the record token' do
+          let(:view_token) { record.view_token }
+          succeed 'the token does not have an expiry set'
+          succeed 'the record has a token expiry set in the future' do
+            before { record.update! view_token_expiry: 1.hour.from_now }
+          end
+          failed 'the record has a token expiry set in the past' do
+            before { record.update! view_token_expiry: 1.hour.ago }
+          end
+        end
+      end
+    end
+  end
+
+  describe_rule :view_marks? do
+    let(:record) { create :complete_result }
+    failed 'role is an instructor' do
+      let(:role) { create(:instructor) }
+    end
+    failed 'role is a ta' do
+      let(:role) { create(:ta) }
+    end
+    failed 'role is a student who is not part of the grouping' do
+      let(:role) { create :student }
+    end
+    describe 'role is a student who is part of the grouping' do
+      let(:role) { create :student }
+      let(:grouping) { create :grouping_with_inviter_and_submission, inviter: role }
+      let(:record) { grouping.current_result }
+      let(:assignment) { record.grouping.assignment }
+      succeed 'assignment.release_with_urls is false' do
+        before { assignment.update! release_with_urls: false }
+      end
+      context 'assignment.release_with_urls is true' do
+        before { assignment.update! release_with_urls: true }
+        let(:context) { { role: role, real_user: role.user, view_token: view_token } }
+        failed 'the view token does not match the record token' do
+          let(:view_token) { "#{record.view_token}abc123" }
+        end
+        context 'the view token matches the record token' do
+          let(:view_token) { record.view_token }
+          succeed 'the token does not have an expiry set'
+          succeed 'the record has a token expiry set in the future' do
+            before { record.update! view_token_expiry: 1.hour.from_now }
+          end
+          failed 'the record has a token expiry set in the past' do
+            before { record.update! view_token_expiry: 1.hour.ago }
+          end
+        end
+      end
     end
   end
 
@@ -274,6 +333,37 @@ describe ResultPolicy do
         end
         failed 'role is not an accepted member of the results grouping' do
           let(:role) { create(:student) }
+        end
+      end
+    end
+  end
+
+  describe_rule :view_token_check? do
+    let(:record) { create :complete_result }
+    failed 'role is an instructor' do
+      let(:role) { create :instructor }
+    end
+    failed 'role is a ta' do
+      let(:role) { create :ta }
+    end
+    failed 'role is a student who is not part of the grouping' do
+      let(:role) { create :student }
+    end
+    describe 'role is a student who is part of the grouping' do
+      let(:role) { create :student }
+      let(:grouping) { create :grouping_with_inviter_and_submission, inviter: role }
+      let(:record) { grouping.current_result }
+      let(:assignment) { record.grouping.assignment }
+      failed 'assignment.release_with_urls is false' do
+        before { assignment.update! release_with_urls: false }
+      end
+      context 'assignment.release_with_urls is true' do
+        before { assignment.update! release_with_urls: true }
+        failed 'view_token is expired' do
+          before { record.update! view_token_expiry: 1.minute.ago }
+        end
+        succeed 'view_token is not expired' do
+          before { record.update! view_token_expiry: 1.minute.from_now }
         end
       end
     end